combine async and sync processing functions in one class
