+++ docs/snippets/batch/gettingStartedErrorHandling.ts
@@ -1,7 +1,7 @@
 import {
   BatchProcessor,
   EventType,
-  processPartialResponse,
+  processPartialResponseSync,
 } from '@aws-lambda-powertools/batch';
 import { Logger } from '@aws-lambda-powertools/logger';
 import type {
@@ -37,7 +37,7 @@ export const handler = async (
   context: Context
 ): Promise<SQSBatchResponse> => {
   // prettier-ignore
-  return processPartialResponse(event, recordHandler, processor, { // (2)!
+  return processPartialResponseSync(event, recordHandler, processor, { // (2)!
     context,
   });
 };
