+++ packages/batch/src/BasePartialProcessor.ts
@@ -35,40 +35,6 @@ abstract class BasePartialProcessor {
   }
 
   /**
-   * Call instance's handler for each record
-   * @returns List of processed records
-   */
-  public async asyncProcess(): Promise<(SuccessResponse | FailureResponse)[]> {
-    /**
-     * If this is a sync processor, user should have called process instead,
-     * so we call the method early to throw the error early thus failing fast.
-     */
-    if (this.constructor.name === 'BatchProcessor') {
-      await this.asyncProcessRecord(this.records[0]);
-    }
-    this.prepare();
-
-    const processingPromises: Promise<SuccessResponse | FailureResponse>[] =
-      this.records.map((record) => this.asyncProcessRecord(record));
-
-    const processedRecords: (SuccessResponse | FailureResponse)[] =
-      await Promise.all(processingPromises);
-
-    this.clean();
-
-    return processedRecords;
-  }
-
-  /**
-   * Process a record with an asyncronous handler
-   *
-   * @param record Record to be processed
-   */
-  public abstract asyncProcessRecord(
-    record: BaseRecord
-  ): Promise<SuccessResponse | FailureResponse>;
-
-  /**
    * Clean class instance after processing
    */
   public abstract clean(): void;
@@ -99,20 +65,14 @@ abstract class BasePartialProcessor {
    * Call instance's handler for each record
    * @returns List of processed records
    */
-  public process(): (SuccessResponse | FailureResponse)[] {
-    /**
-     * If this is an async processor, user should have called processAsync instead,
-     * so we call the method early to throw the error early thus failing fast.
-     */
-    if (this.constructor.name === 'AsyncBatchProcessor') {
-      this.processRecord(this.records[0]);
-    }
+  public async process(): Promise<(SuccessResponse | FailureResponse)[]> {
     this.prepare();
 
-    const processedRecords: (SuccessResponse | FailureResponse)[] = [];
-    for (const record of this.records) {
-      processedRecords.push(this.processRecord(record));
-    }
+    const processingPromises: Promise<SuccessResponse | FailureResponse>[] =
+      this.records.map((record) => this.processRecord(record));
+
+    const processedRecords: (SuccessResponse | FailureResponse)[] =
+      await Promise.all(processingPromises);
 
     this.clean();
 
@@ -120,12 +80,38 @@ abstract class BasePartialProcessor {
   }
 
   /**
-   * Process a record with the handler
+   * Process a record with an asyncronous handler
+   *
    * @param record Record to be processed
    */
   public abstract processRecord(
     record: BaseRecord
+  ): Promise<SuccessResponse | FailureResponse>;
+
+  /**
+   * Process a record with the handler
+   * @param record Record to be processed
+   */
+  public abstract processRecordSync(
+    record: BaseRecord
   ): SuccessResponse | FailureResponse;
+
+  /**
+   * Call instance's handler for each record
+   * @returns List of processed records
+   */
+  public processSync(): (SuccessResponse | FailureResponse)[] {
+    this.prepare();
+
+    const processedRecords: (SuccessResponse | FailureResponse)[] = [];
+    for (const record of this.records) {
+      processedRecords.push(this.processRecordSync(record));
+    }
+
+    this.clean();
+
+    return processedRecords;
+  }
 
   /**
    * Set class instance attributes before execution
