--- packages/batch/src/asyncProcessPartialResponse.ts
@@ -1,34 +1,0 @@
-import { BasePartialBatchProcessor } from './BasePartialBatchProcessor';
-import { UnexpectedBatchTypeError } from './errors';
-import type {
-  BaseRecord,
-  BatchProcessingOptions,
-  PartialItemFailureResponse,
-} from './types';
-
-/**
- * Higher level function to handle batch event processing
- * @param event Lambda's original event
- * @param recordHandler Callable function to process each record from the batch
- * @param processor Batch processor to handle partial failure cases
- * @param options Batch processing options
- * @returns Lambda Partial Batch Response
- */
-const asyncProcessPartialResponse = async (
-  event: { Records: BaseRecord[] },
-  recordHandler: CallableFunction,
-  processor: BasePartialBatchProcessor,
-  options?: BatchProcessingOptions
-): Promise<PartialItemFailureResponse> => {
-  if (!event.Records || !Array.isArray(event.Records)) {
-    throw new UnexpectedBatchTypeError();
-  }
-
-  processor.register(event.Records, recordHandler, options);
-
-  await processor.asyncProcess();
-
-  return processor.response();
-};
-
-export { asyncProcessPartialResponse };
