+++ packages/batch/src/index.ts
@@ -2,9 +2,6 @@ export * from './constants';
 export * from './errors';
 export * from './types';
 export * from './BasePartialProcessor';
-export * from './BasePartialBatchProcessor';
 export * from './BatchProcessor';
-export * from './AsyncBatchProcessor';
 export * from './processPartialResponse';
-export * from './asyncProcessPartialResponse';
 export * from './SqsFifoPartialProcessor';
