+++ packages/batch/src/processPartialResponse.ts
@@ -1,10 +1,10 @@
-import { BasePartialBatchProcessor } from './BasePartialBatchProcessor';
 import { UnexpectedBatchTypeError } from './errors';
 import type {
   BaseRecord,
   BatchProcessingOptions,
   PartialItemFailureResponse,
 } from './types';
+import { BatchProcessor } from './BatchProcessor';
 
 /**
  * Higher level function to handle batch event processing
@@ -14,10 +14,35 @@ import type {
  * @param options Batch processing options
  * @returns Lambda Partial Batch Response
  */
-const processPartialResponse = (
+const processPartialResponse = async (
   event: { Records: BaseRecord[] },
   recordHandler: CallableFunction,
-  processor: BasePartialBatchProcessor,
+  processor: BatchProcessor,
+  options?: BatchProcessingOptions
+): Promise<PartialItemFailureResponse> => {
+  if (!event.Records || !Array.isArray(event.Records)) {
+    throw new UnexpectedBatchTypeError();
+  }
+
+  processor.register(event.Records, recordHandler, options);
+
+  await processor.process();
+
+  return processor.response();
+};
+
+/**
+ * Higher level function to handle batch event processing
+ * @param event Lambda's original event
+ * @param recordHandler Callable function to process each record from the batch
+ * @param processor Batch processor to handle partial failure cases
+ * @param options Batch processing options
+ * @returns Lambda Partial Batch Response
+ */
+const processPartialResponseSync = (
+  event: { Records: BaseRecord[] },
+  recordHandler: CallableFunction,
+  processor: BatchProcessor,
   options?: BatchProcessingOptions
 ): PartialItemFailureResponse => {
   if (!event.Records || !Array.isArray(event.Records)) {
@@ -26,9 +51,9 @@ const processPartialResponse = (
 
   processor.register(event.Records, recordHandler, options);
 
-  processor.process();
+  processor.processSync();
 
   return processor.response();
 };
 
-export { processPartialResponse };
+export { processPartialResponse, processPartialResponseSync };
