+++ packages/batch/tests/unit/BatchProcessor.test.ts
@@ -5,9 +5,8 @@
  */
 import type { Context } from 'aws-lambda';
 import { helloworldContext as dummyContext } from '@aws-lambda-powertools/commons/lib/samples/resources/contexts';
-import { BatchProcessor } from '../../src/BatchProcessor';
 import { EventType } from '../../src/constants';
-import { BatchProcessingError, FullBatchFailureError } from '../../src/errors';
+import { FullBatchFailureError } from '../../src/errors';
 import type { BatchProcessingOptions } from '../../src/types';
 import {
   dynamodbRecordFactory,
@@ -15,11 +14,16 @@ import {
   sqsRecordFactory,
 } from '../helpers/factories';
 import {
+  asyncDynamodbRecordHandler,
+  asyncHandlerWithContext,
+  asyncKinesisRecordHandler,
+  asyncSqsRecordHandler,
   dynamodbRecordHandler,
   handlerWithContext,
   kinesisRecordHandler,
   sqsRecordHandler,
 } from '../helpers/handlers';
+import { BatchProcessor } from '../../src/BatchProcessor';
 
 describe('Class: BatchProcessor', () => {
   const ENVIRONMENT_VARIABLES = process.env;
@@ -45,7 +49,7 @@ describe('Class: BatchProcessor', () => {
 
       // Act
       processor.register(records, sqsRecordHandler);
-      const processedMessages = processor.process();
+      const processedMessages = processor.processSync();
 
       // Assess
       expect(processedMessages).toStrictEqual([
@@ -64,7 +68,7 @@ describe('Class: BatchProcessor', () => {
 
       // Act
       processor.register(records, sqsRecordHandler);
-      const processedMessages = processor.process();
+      const processedMessages = processor.processSync();
 
       // Assess
       expect(processedMessages[1]).toStrictEqual([
@@ -92,7 +96,7 @@ describe('Class: BatchProcessor', () => {
 
       // Act & Assess
       processor.register(records, sqsRecordHandler);
-      expect(() => processor.process()).toThrowError(FullBatchFailureError);
+      expect(() => processor.processSync()).toThrowError(FullBatchFailureError);
     });
   });
 
@@ -106,7 +110,7 @@ describe('Class: BatchProcessor', () => {
 
       // Act
       processor.register(records, kinesisRecordHandler);
-      const processedMessages = processor.process();
+      const processedMessages = processor.processSync();
 
       // Assess
       expect(processedMessages).toStrictEqual([
@@ -125,7 +129,7 @@ describe('Class: BatchProcessor', () => {
 
       // Act
       processor.register(records, kinesisRecordHandler);
-      const processedMessages = processor.process();
+      const processedMessages = processor.processSync();
 
       // Assess
       expect(processedMessages[1]).toStrictEqual([
@@ -154,7 +158,7 @@ describe('Class: BatchProcessor', () => {
       processor.register(records, kinesisRecordHandler);
 
       // Assess
-      expect(() => processor.process()).toThrowError(FullBatchFailureError);
+      expect(() => processor.processSync()).toThrowError(FullBatchFailureError);
     });
   });
 
@@ -168,7 +172,7 @@ describe('Class: BatchProcessor', () => {
 
       // Act
       processor.register(records, dynamodbRecordHandler);
-      const processedMessages = processor.process();
+      const processedMessages = processor.processSync();
 
       // Assess
       expect(processedMessages).toStrictEqual([
@@ -187,7 +191,7 @@ describe('Class: BatchProcessor', () => {
 
       // Act
       processor.register(records, dynamodbRecordHandler);
-      const processedMessages = processor.process();
+      const processedMessages = processor.processSync();
 
       // Assess
       expect(processedMessages[1]).toStrictEqual([
@@ -217,7 +221,7 @@ describe('Class: BatchProcessor', () => {
       processor.register(records, dynamodbRecordHandler);
 
       // Assess
-      expect(() => processor.process()).toThrowError(FullBatchFailureError);
+      expect(() => processor.processSync()).toThrowError(FullBatchFailureError);
     });
   });
 
@@ -231,7 +235,7 @@ describe('Class: BatchProcessor', () => {
 
       // Act
       processor.register(records, handlerWithContext, options);
-      const processedMessages = processor.process();
+      const processedMessages = processor.processSync();
 
       // Assess
       expect(processedMessages).toStrictEqual([
@@ -249,7 +253,7 @@ describe('Class: BatchProcessor', () => {
 
       // Act
       processor.register(records, sqsRecordHandler, options);
-      const processedMessages = processor.process();
+      const processedMessages = processor.processSync();
 
       // Assess
       expect(processedMessages).toStrictEqual([
@@ -269,17 +273,256 @@ describe('Class: BatchProcessor', () => {
 
       // Act
       processor.register(records, handlerWithContext, badOptions);
-      expect(() => processor.process()).toThrowError(FullBatchFailureError);
+      expect(() => processor.processSync()).toThrowError(FullBatchFailureError);
     });
   });
 
-  test('When calling the async process method, it should throw an error', async () => {
-    // Prepare
-    const processor = new BatchProcessor(EventType.SQS);
+  describe('Asynchronously processing SQS Records', () => {
+    test('Batch processing SQS records with no failures', async () => {
+      // Prepare
+      const firstRecord = sqsRecordFactory('success');
+      const secondRecord = sqsRecordFactory('success');
+      const records = [firstRecord, secondRecord];
+      const processor = new BatchProcessor(EventType.SQS);
 
-    // Act & Assess
-    await expect(() => processor.asyncProcess()).rejects.toThrowError(
-      BatchProcessingError
-    );
+      // Act
+      processor.register(records, asyncSqsRecordHandler);
+      const processedMessages = await processor.process();
+
+      // Assess
+      expect(processedMessages).toStrictEqual([
+        ['success', firstRecord.body, firstRecord],
+        ['success', secondRecord.body, secondRecord],
+      ]);
+    });
+
+    test('Batch processing SQS records with some failures', async () => {
+      // Prepare
+      const firstRecord = sqsRecordFactory('failure');
+      const secondRecord = sqsRecordFactory('success');
+      const thirdRecord = sqsRecordFactory('fail');
+      const records = [firstRecord, secondRecord, thirdRecord];
+      const processor = new BatchProcessor(EventType.SQS);
+
+      // Act
+      processor.register(records, asyncSqsRecordHandler);
+      const processedMessages = await processor.process();
+
+      // Assess
+      expect(processedMessages[1]).toStrictEqual([
+        'success',
+        secondRecord.body,
+        secondRecord,
+      ]);
+      expect(processor.failureMessages.length).toBe(2);
+      expect(processor.response()).toStrictEqual({
+        batchItemFailures: [
+          { itemIdentifier: firstRecord.messageId },
+          { itemIdentifier: thirdRecord.messageId },
+        ],
+      });
+    });
+
+    test('Batch processing SQS records with all failures', async () => {
+      // Prepare
+      const firstRecord = sqsRecordFactory('failure');
+      const secondRecord = sqsRecordFactory('failure');
+      const thirdRecord = sqsRecordFactory('fail');
+
+      const records = [firstRecord, secondRecord, thirdRecord];
+      const processor = new BatchProcessor(EventType.SQS);
+
+      // Act
+      processor.register(records, asyncSqsRecordHandler);
+
+      // Assess
+      await expect(processor.process()).rejects.toThrowError(
+        FullBatchFailureError
+      );
+    });
+  });
+
+  describe('Asynchronously processing Kinesis Records', () => {
+    test('Batch processing Kinesis records with no failures', async () => {
+      // Prepare
+      const firstRecord = kinesisRecordFactory('success');
+      const secondRecord = kinesisRecordFactory('success');
+      const records = [firstRecord, secondRecord];
+      const processor = new BatchProcessor(EventType.KinesisDataStreams);
+
+      // Act
+      processor.register(records, asyncKinesisRecordHandler);
+      const processedMessages = await processor.process();
+
+      // Assess
+      expect(processedMessages).toStrictEqual([
+        ['success', firstRecord.kinesis.data, firstRecord],
+        ['success', secondRecord.kinesis.data, secondRecord],
+      ]);
+    });
+
+    test('Batch processing Kinesis records with some failures', async () => {
+      // Prepare
+      const firstRecord = kinesisRecordFactory('failure');
+      const secondRecord = kinesisRecordFactory('success');
+      const thirdRecord = kinesisRecordFactory('fail');
+      const records = [firstRecord, secondRecord, thirdRecord];
+      const processor = new BatchProcessor(EventType.KinesisDataStreams);
+
+      // Act
+      processor.register(records, asyncKinesisRecordHandler);
+      const processedMessages = await processor.process();
+
+      // Assess
+      expect(processedMessages[1]).toStrictEqual([
+        'success',
+        secondRecord.kinesis.data,
+        secondRecord,
+      ]);
+      expect(processor.failureMessages.length).toBe(2);
+      expect(processor.response()).toStrictEqual({
+        batchItemFailures: [
+          { itemIdentifier: firstRecord.kinesis.sequenceNumber },
+          { itemIdentifier: thirdRecord.kinesis.sequenceNumber },
+        ],
+      });
+    });
+
+    test('Batch processing Kinesis records with all failures', async () => {
+      // Prepare
+      const firstRecord = kinesisRecordFactory('failure');
+      const secondRecord = kinesisRecordFactory('failure');
+      const thirdRecord = kinesisRecordFactory('fail');
+
+      const records = [firstRecord, secondRecord, thirdRecord];
+      const processor = new BatchProcessor(EventType.KinesisDataStreams);
+
+      // Act
+      processor.register(records, asyncKinesisRecordHandler);
+
+      // Assess
+      await expect(processor.process()).rejects.toThrowError(
+        FullBatchFailureError
+      );
+    });
+  });
+
+  describe('Asynchronously processing DynamoDB Records', () => {
+    test('Batch processing DynamoDB records with no failures', async () => {
+      // Prepare
+      const firstRecord = dynamodbRecordFactory('success');
+      const secondRecord = dynamodbRecordFactory('success');
+      const records = [firstRecord, secondRecord];
+      const processor = new BatchProcessor(EventType.DynamoDBStreams);
+
+      // Act
+      processor.register(records, asyncDynamodbRecordHandler);
+      const processedMessages = await processor.process();
+
+      // Assess
+      expect(processedMessages).toStrictEqual([
+        ['success', firstRecord.dynamodb?.NewImage?.Message, firstRecord],
+        ['success', secondRecord.dynamodb?.NewImage?.Message, secondRecord],
+      ]);
+    });
+
+    test('Batch processing DynamoDB records with some failures', async () => {
+      // Prepare
+      const firstRecord = dynamodbRecordFactory('failure');
+      const secondRecord = dynamodbRecordFactory('success');
+      const thirdRecord = dynamodbRecordFactory('fail');
+      const records = [firstRecord, secondRecord, thirdRecord];
+      const processor = new BatchProcessor(EventType.DynamoDBStreams);
+
+      // Act
+      processor.register(records, asyncDynamodbRecordHandler);
+      const processedMessages = await processor.process();
+
+      // Assess
+      expect(processedMessages[1]).toStrictEqual([
+        'success',
+        secondRecord.dynamodb?.NewImage?.Message,
+        secondRecord,
+      ]);
+      expect(processor.failureMessages.length).toBe(2);
+      expect(processor.response()).toStrictEqual({
+        batchItemFailures: [
+          { itemIdentifier: firstRecord.dynamodb?.SequenceNumber },
+          { itemIdentifier: thirdRecord.dynamodb?.SequenceNumber },
+        ],
+      });
+    });
+
+    test('Batch processing DynamoDB records with all failures', async () => {
+      // Prepare
+      const firstRecord = dynamodbRecordFactory('failure');
+      const secondRecord = dynamodbRecordFactory('failure');
+      const thirdRecord = dynamodbRecordFactory('fail');
+
+      const records = [firstRecord, secondRecord, thirdRecord];
+      const processor = new BatchProcessor(EventType.DynamoDBStreams);
+
+      // Act
+      processor.register(records, asyncDynamodbRecordHandler);
+
+      // Assess
+      await expect(processor.process()).rejects.toThrowError(
+        FullBatchFailureError
+      );
+    });
+  });
+
+  describe('Batch processing with Lambda context', () => {
+    test('Batch processing when context is provided and handler accepts', async () => {
+      // Prepare
+      const firstRecord = sqsRecordFactory('success');
+      const secondRecord = sqsRecordFactory('success');
+      const records = [firstRecord, secondRecord];
+      const processor = new BatchProcessor(EventType.SQS);
+
+      // Act
+      processor.register(records, asyncHandlerWithContext, options);
+      const processedMessages = await processor.process();
+
+      // Assess
+      expect(processedMessages).toStrictEqual([
+        ['success', firstRecord.body, firstRecord],
+        ['success', secondRecord.body, secondRecord],
+      ]);
+    });
+
+    test('Batch processing when context is provided and handler does not accept', async () => {
+      // Prepare
+      const firstRecord = sqsRecordFactory('success');
+      const secondRecord = sqsRecordFactory('success');
+      const records = [firstRecord, secondRecord];
+      const processor = new BatchProcessor(EventType.SQS);
+
+      // Act
+      processor.register(records, asyncSqsRecordHandler, options);
+      const processedMessages = await processor.process();
+
+      // Assess
+      expect(processedMessages).toStrictEqual([
+        ['success', firstRecord.body, firstRecord],
+        ['success', secondRecord.body, secondRecord],
+      ]);
+    });
+
+    test('Batch processing when malformed context is provided and handler attempts to use', async () => {
+      // Prepare
+      const firstRecord = sqsRecordFactory('success');
+      const secondRecord = sqsRecordFactory('success');
+      const records = [firstRecord, secondRecord];
+      const processor = new BatchProcessor(EventType.SQS);
+      const badContext = { foo: 'bar' };
+      const badOptions = { context: badContext as unknown as Context };
+
+      // Act
+      processor.register(records, asyncHandlerWithContext, badOptions);
+      await expect(() => processor.process()).rejects.toThrowError(
+        FullBatchFailureError
+      );
+    });
   });
 });
