+++ packages/batch/tests/unit/processPartialResponse.test.ts
@@ -11,7 +11,6 @@ import type {
 } from 'aws-lambda';
 import { helloworldContext as dummyContext } from '@aws-lambda-powertools/commons/lib/samples/resources/contexts';
 import { Custom as dummyEvent } from '@aws-lambda-powertools/commons/lib/samples/resources/events';
-import { BatchProcessor, processPartialResponse } from '../../src';
 import { EventType } from '../../src/constants';
 import type {
   BatchProcessingOptions,
@@ -23,11 +22,20 @@ import {
   sqsRecordFactory,
 } from '../helpers/factories';
 import {
+  asyncDynamodbRecordHandler,
+  asyncHandlerWithContext,
+  asyncKinesisRecordHandler,
+  asyncSqsRecordHandler,
   dynamodbRecordHandler,
   handlerWithContext,
   kinesisRecordHandler,
   sqsRecordHandler,
 } from '../helpers/handlers';
+import { BatchProcessor } from '../../src/BatchProcessor';
+import {
+  processPartialResponse,
+  processPartialResponseSync,
+} from '../../src/processPartialResponse';
 
 describe('Function: processPartialResponse()', () => {
   const ENVIRONMENT_VARIABLES = process.env;
@@ -55,7 +63,11 @@ describe('Function: processPartialResponse()', () => {
       const processor = new BatchProcessor(EventType.SQS);
 
       // Act
-      const ret = processPartialResponse(batch, sqsRecordHandler, processor);
+      const ret = processPartialResponseSync(
+        batch,
+        sqsRecordHandler,
+        processor
+      );
 
       // Assess
       expect(ret).toStrictEqual({ batchItemFailures: [] });
@@ -71,9 +83,50 @@ describe('Function: processPartialResponse()', () => {
       const processor = new BatchProcessor(EventType.SQS);
 
       // Act
-      const ret = processPartialResponse(
+      const ret = processPartialResponseSync(
         batch,
         handlerWithContext,
+        processor,
+        options
+      );
+
+      // Assess
+      expect(ret).toStrictEqual({ batchItemFailures: [] });
+    });
+
+    test('Process partial response function call with asynchronous handler', async () => {
+      // Prepare
+      const records = [
+        sqsRecordFactory('success'),
+        sqsRecordFactory('success'),
+      ];
+      const batch = { Records: records };
+      const processor = new BatchProcessor(EventType.SQS);
+
+      // Act
+      const ret = await processPartialResponse(
+        batch,
+        asyncSqsRecordHandler,
+        processor
+      );
+
+      // Assess
+      expect(ret).toStrictEqual({ batchItemFailures: [] });
+    });
+
+    test('Process partial response function call with context provided', async () => {
+      // Prepare
+      const records = [
+        sqsRecordFactory('success'),
+        sqsRecordFactory('success'),
+      ];
+      const batch = { Records: records };
+      const processor = new BatchProcessor(EventType.SQS);
+
+      // Act
+      const ret = await processPartialResponse(
+        batch,
+        asyncHandlerWithContext,
         processor,
         options
       );
@@ -97,7 +150,7 @@ describe('Function: processPartialResponse()', () => {
         event: SQSEvent,
         _context: Context
       ): PartialItemFailureResponse => {
-        return processPartialResponse(event, sqsRecordHandler, processor);
+        return processPartialResponseSync(event, sqsRecordHandler, processor);
       };
 
       // Act
@@ -120,7 +173,11 @@ describe('Function: processPartialResponse()', () => {
         event: KinesisStreamEvent,
         _context: Context
       ): PartialItemFailureResponse => {
-        return processPartialResponse(event, kinesisRecordHandler, processor);
+        return processPartialResponseSync(
+          event,
+          kinesisRecordHandler,
+          processor
+        );
       };
 
       // Act
@@ -143,7 +200,11 @@ describe('Function: processPartialResponse()', () => {
         event: DynamoDBStreamEvent,
         _context: Context
       ): PartialItemFailureResponse => {
-        return processPartialResponse(event, dynamodbRecordHandler, processor);
+        return processPartialResponseSync(
+          event,
+          dynamodbRecordHandler,
+          processor
+        );
       };
 
       // Act
@@ -162,7 +223,7 @@ describe('Function: processPartialResponse()', () => {
         event: SQSEvent,
         _context: Context
       ): PartialItemFailureResponse => {
-        return processPartialResponse(event, sqsRecordHandler, processor);
+        return processPartialResponseSync(event, sqsRecordHandler, processor);
       };
 
       // Act & Assess
@@ -188,7 +249,7 @@ describe('Function: processPartialResponse()', () => {
       ): PartialItemFailureResponse => {
         const options: BatchProcessingOptions = { context: context };
 
-        return processPartialResponse(
+        return processPartialResponseSync(
           event,
           handlerWithContext,
           processor,
@@ -198,6 +259,139 @@ describe('Function: processPartialResponse()', () => {
 
       // Act
       const result = handler(event, context);
+
+      // Assess
+      expect(result).toStrictEqual({ batchItemFailures: [] });
+    });
+
+    test('Process partial response through handler with SQS event', async () => {
+      // Prepare
+      const records = [
+        sqsRecordFactory('success'),
+        sqsRecordFactory('success'),
+      ];
+      const processor = new BatchProcessor(EventType.SQS);
+      const event: SQSEvent = { Records: records };
+
+      const handler = async (
+        event: SQSEvent,
+        _context: Context
+      ): Promise<PartialItemFailureResponse> => {
+        return processPartialResponse(event, asyncSqsRecordHandler, processor);
+      };
+
+      // Act
+      const result = await handler(event, context);
+
+      // Assess
+      expect(result).toStrictEqual({ batchItemFailures: [] });
+    });
+
+    test('Process partial response through handler with Kinesis event', async () => {
+      // Prepare
+      const records = [
+        kinesisRecordFactory('success'),
+        kinesisRecordFactory('success'),
+      ];
+      const processor = new BatchProcessor(EventType.KinesisDataStreams);
+      const event: KinesisStreamEvent = { Records: records };
+
+      const handler = async (
+        event: KinesisStreamEvent,
+        _context: Context
+      ): Promise<PartialItemFailureResponse> => {
+        return await processPartialResponse(
+          event,
+          asyncKinesisRecordHandler,
+          processor
+        );
+      };
+
+      // Act
+      const result = await handler(event, context);
+
+      // Assess
+      expect(result).toStrictEqual({ batchItemFailures: [] });
+    });
+
+    test('Process partial response through handler with DynamoDB event', async () => {
+      // Prepare
+      const records = [
+        dynamodbRecordFactory('success'),
+        dynamodbRecordFactory('success'),
+      ];
+      const processor = new BatchProcessor(EventType.DynamoDBStreams);
+      const event: DynamoDBStreamEvent = { Records: records };
+
+      const handler = async (
+        event: DynamoDBStreamEvent,
+        _context: Context
+      ): Promise<PartialItemFailureResponse> => {
+        return await processPartialResponse(
+          event,
+          asyncDynamodbRecordHandler,
+          processor
+        );
+      };
+
+      // Act
+      const result = await handler(event, context);
+
+      // Assess
+      expect(result).toStrictEqual({ batchItemFailures: [] });
+    });
+
+    test('Process partial response through handler for SQS records with incorrect event type', async () => {
+      // Prepare
+      const processor = new BatchProcessor(EventType.SQS);
+      const event = dummyEvent;
+
+      const handler = async (
+        event: SQSEvent,
+        _context: Context
+      ): Promise<PartialItemFailureResponse> => {
+        return await processPartialResponse(
+          event,
+          asyncSqsRecordHandler,
+          processor
+        );
+      };
+
+      // Act & Assess
+      await expect(() =>
+        handler(event as unknown as SQSEvent, context)
+      ).rejects.toThrowError(
+        `Unexpected batch type. Possible values are: ${Object.keys(
+          EventType
+        ).join(', ')}`
+      );
+    });
+
+    test('Process partial response through handler with context provided', async () => {
+      // Prepare
+      const records = [
+        sqsRecordFactory('success'),
+        sqsRecordFactory('success'),
+      ];
+      const processor = new BatchProcessor(EventType.SQS);
+      const event: SQSEvent = { Records: records };
+
+      const handler = async (
+        event: SQSEvent,
+        context: Context
+      ): Promise<PartialItemFailureResponse> => {
+        const options: BatchProcessingOptions = { context: context };
+
+        return await processPartialResponse(
+          event,
+          asyncHandlerWithContext,
+          processor,
+          options
+        );
+      };
+
+      // Act
+      const result = await handler(event, context);
 
       // Assess
       expect(result).toStrictEqual({ batchItemFailures: [] });
